migrate test content to separate directory
