+++ packages/ts-sdk/src/test/__fixtures__/from_root_examples.ts
