+++ packages/ts-sdk/src/test/__fixtures__/small_quadstore.ts
